requirements.txt
aiogram>=3.3.0
python-dotenv>=1.0.0
Pillow>=10.2.0
opencv-python>=4.8.0
pydantic==2.5.3 